xử lý ajax tăng giảm sản phẩm, đổi decimal price thành double trong models,(đổi trong db sau), đổi giá VNĐ thành $ ,

+++ SanThuongMaiG15/ModelViews/CartItem.cs
@@ -6,6 +6,6 @@ namespace SanThuongMaiG15.ModelViews
     {
         public Product product { get; set; }    
         public int quantity { get; set; }
-        public decimal TotalMoney => quantity * (product?.Price ?? 0); 
+        public double TotalMoney => quantity * (product?.Price ?? 0); 
     }
 }
